Create phantoms module for the fish
