+++ tests/test_transmit_schemes.py
@@ -6,6 +6,7 @@ import pytest
 
 from zea import ops
 from zea.beamform.delays import compute_t0_delays_focused, compute_t0_delays_planewave
+from zea.beamform.phantoms import fish
 from zea.probes import Probe
 from zea.scan import Scan
 
@@ -86,48 +87,6 @@ def default_pipeline():
     pipeline.prepend(ops.Simulate())
     pipeline.append(ops.Normalize(input_range=ops.DEFAULT_DYNAMIC_RANGE, output_range=(0, 255)))
     return pipeline
-
-
-def _get_fish_phantom():
-    """Returns a scatterer phantom for ultrasound simulation tests.
-
-    Returns:
-        ndarray: The scatterer positions of shape (n_scat, 3).
-    """
-    # The size is the height of the fish
-    size = 11e-3
-    z_offset = 2.0 * size
-
-    # See https://en.wikipedia.org/wiki/Fish_curve
-    def fish_curve(t, size=1):
-        x = size * (np.cos(t) - np.sin(t) ** 2 / np.sqrt(2))
-        y = size * np.cos(t) * np.sin(t)
-        return x, y
-
-    scat_x, scat_z = fish_curve(np.linspace(0, 2 * np.pi, 100), size=size)
-
-    scat_x = np.concatenate(
-        [
-            scat_x,
-            np.array([size * 0.7]),
-            np.array([size * 1.1]),
-            np.array([size * 1.4]),
-            np.array([size * 1.2]),
-        ]
-    )
-    scat_y = np.zeros_like(scat_x)
-    scat_z = np.concatenate(
-        [
-            scat_z,
-            np.array([-size * 0.1]),
-            np.array([-size * 0.25]),
-            np.array([-size * 0.6]),
-            np.array([-size * 1.0]),
-        ]
-    )
-
-    scat_z += z_offset
-    return np.stack([scat_x, scat_y, scat_z], axis=1)
 
 
 def _get_linear_probe():
@@ -460,7 +419,7 @@ def _test_location(image, extent, true_position):
 @pytest.fixture
 def ultrasound_scatterers():
     """Returns scatterer positions and magnitudes for ultrasound simulation tests."""
-    scat_positions = _get_fish_phantom()
+    scat_positions = fish()
     n_scat = scat_positions.shape[0]
 
     return {
